% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC}
\alias{AUC}
\title{Area Under the Curve (AUC)}
\usage{
AUC(dta, Time = "Time", Conc = "Conc", method = "Linear-Up/Log-Down")
}
\arguments{
\item{dta}{Dataframe with concentration-time data.}

\item{Time}{Name of the column with time data.}

\item{Conc}{Name of the column with concentration data.}

\item{method}{Method for the AUC calculation:
\itemize{
\item \code{Linear-Up/Log-Down} (default method): Linear Trapezoidal Rule for for Increasing Values, Log Trapezoidal Rule for Decreasing Values.
\item \code{Linear/Log}: Linear Trapezoidal Rule for for Values below Cmax, Log Trapezoidal Rule for Values above Cmax.
\item \code{Linear}: Linear Trapezoidal Rule.
}}
}
\value{
Dataframe with cumulative AUC over time.
}
\description{
\code{AUC()} calculates the cumulative area under the concentration-time curve (AUC) over time.
}
\examples{
dta <- data.frame(
  Time = c(0, 0.25, 0.5, 0.75, 1, 1.5, 1.75, 2, 2.25, 2.5,
           2.75, 3, 3.25, 3.5, 3.75, 4, 6, 8, 12, 24),
  Conc = c(0.00, 221.23, 377.19, 494.73, 555.74, 623.86, 615.45, 663.38, 660.29, 621.71,
           650.33, 622.28, 626.72, 574.94, 610.51, 554.02, 409.14, 299.76, 162.85, 27.01))
AUC(dta, Time = 'Time', Conc = 'Conc')

}
