% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cmax.R
\name{Cmax}
\alias{Cmax}
\title{Maximum Concentration (Cmax)}
\usage{
Cmax(dta, Conc = "Conc")
}
\arguments{
\item{dta}{Dataframe with concentration-time data.}

\item{Conc}{Name of the column with concentration data.}
}
\value{
Value of the maximum observed concentration.
}
\description{
\code{Cmax()} calculates the maximum observed concentration post-dose \eqn{C_{\text{max}}}
directly obtained from the observed concentration-time profile.
}
\examples{
dta <- data.frame(
  Time = c(0, 0.25, 0.5, 0.75, 1, 1.5, 1.75, 2, 2.25, 2.5,
           2.75, 3, 3.25, 3.5, 3.75, 4, 6, 8, 12, 24),
  Conc = c(0.00, 221.23, 377.19, 494.73, 555.74, 623.86, 615.45, 663.38, 660.29, 621.71,
           650.33, 622.28, 626.72, 574.94, 610.51, 554.02, 409.14, 299.76, 162.85, 27.01))
Cmax(dta, 'Conc')

}
