% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{Geometric Mean}
\usage{
geomean(x, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be stripped
before the computation proceeds. Defaults to \code{FALSE}.}
}
\value{
Value of the geometric mean.
}
\description{
\code{geomean()} calculates geometric mean, i.e., the \emph{N}th root of the product of the \emph{N}
observations, equivalent to exp(mean(log(x))).
}
\examples{
geomean(x, TRUE)
geomean(x, FALSE)

# Can be used for the calculation of geometric mean of concentration data,
# for Test and Reference treatments, over time.
# Calculate the geometric mean for each Treatment, by timepoint
dta <- data.frame(t(tapply(dta_id$Conc,
                           dta_id[,c('Treatment','Time')],
                           geomean)))
dta$Time <- as.numeric(row.names(dta))
row.names(dta) <- c()

# To stack treatment information
dta <- cbind(dta[3],stack(dta[1:2]))
names(dta)[names(dta) \%in\% c('values','ind')] <- c('Conc',
                                                   'Treatment')

}
\author{
Sara Carolina Henriques
}
