% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tmax.R
\name{Tmax}
\alias{Tmax}
\title{Time of Cmax (Tmax)}
\usage{
Tmax(dta, Time = "Time", Conc = "Conc")
}
\arguments{
\item{dta}{Dataframe with concentration-time data.}

\item{Time}{Name of the column with time data.}

\item{Conc}{Name of the column with concentration data.}
}
\value{
Value of the time of the maximum observed concentration.
}
\description{
\code{Tmax()} calculates the time of the maximum observed concentration post-dose (\eqn{t_{\text{max}}}).
}
\examples{
Tmax(dta, 't', 'Conc')
Tmax(x, 'Time', 'Concentration')

}
