% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC}
\alias{AUC}
\title{Area Under the Curve (AUC)}
\usage{
AUC(dta, Time = "Time", Conc = "Conc", method = "Linear-Up/Log-Down")
}
\arguments{
\item{dta}{Dataframe with concentration-time data.}

\item{Time}{Name of the column with time data.}

\item{Conc}{Name of the column with concentration data.}

\item{method}{Method for the AUC calculation:
\itemize{
\item \code{Linear-Up/Log-Down} (default method): Linear Trapezoidal Rule for for Increasing Values, Log Trapezoidal Rule for Decreasing Values.
\item \code{Linear/Log}: Linear Trapezoidal Rule for for Values below Cmax, Log Trapezoidal Rule for Values above Cmax.
\item \code{Linear}: Linear Trapezoidal Rule.
}}
}
\value{
Dataframe with cumulative AUC over time.
}
\description{
\code{AUC()} calculates the cumulative area under the concentration-time curve (AUC) over time.
}
\examples{
AUC(dta, 't', 'Conc')
AUC(x, 'Time', 'Concentration', 'Linear/Log')

}
\author{
Sara Carolina Henriques
}
