% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUClast.R
\name{AUClast}
\alias{AUClast}
\title{AUC to Last Timepoint}
\usage{
AUClast(dta, Time = "Time", Conc = "Conc", method = "Linear-Up/Log-Down")
}
\arguments{
\item{dta}{Dataframe with concentration-time data.}

\item{Time}{Name of the column with time data.}

\item{Conc}{Name of the column with concentration data.}

\item{method}{Method for the AUC calculation:
\itemize{
\item \code{Linear-Up/Log-Down} (default method): Linear Trapezoidal Rule for for Increasing Values, Log Trapezoidal Rule for Decreasing Values.
\item \code{Linear/Log}: Linear Trapezoidal Rule for for Values below \eqn{C_{\text{max}}}, Log Trapezoidal Rule for Values above \eqn{C_{\text{max}}}.
\item \code{Linear}: Linear Trapezoidal Rule.
}}
}
\value{
Value of \eqn{AUC_{\text{0-t}}}.
}
\description{
\code{AUClast()} calculates area under the concentration-time curve (AUC)
from time zero to the time of the last measurable concentration (\eqn{AUC_{\text{0-t}}}).
}
\examples{
dta <- data.frame(
  Time = c(0, 0.25, 0.5, 0.75, 1, 1.5, 1.75, 2, 2.25, 2.5,
           2.75, 3, 3.25, 3.5, 3.75, 4, 6, 8, 12, 24),
  Conc = c(0.00, 221.23, 377.19, 494.73, 555.74, 623.86, 615.45, 663.38, 660.29, 621.71,
           650.33, 622.28, 626.72, 574.94, 610.51, 554.02, 409.14, 299.76, 162.85, 27.01))
AUClast(dta, 't', 'Conc')

}
