% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cmax.R
\name{Cmax}
\alias{Cmax}
\title{Maximum Concentration (Cmax)}
\usage{
Cmax(dta, Conc = "Conc")
}
\arguments{
\item{dta}{Dataframe with concentration-time data.}

\item{Conc}{Name of the column with concentration data.}
}
\value{
Value of the maximum observed concentration.
}
\description{
\code{Cmax()} calculates the maximum observed concentration post-dose \eqn{C_{\text{max}}}
directly obtained from the observed concentration-time profile.
}
\examples{
Cmax(dta, 'Conc')
Cmax(x, 'Concentration')

}
\author{
Sara Carolina Henriques
}
