% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2.AUC.R
\name{f2.AUC}
\alias{f2.AUC}
\title{AUC Similarity f2 Factor}
\usage{
f2.AUC(
  dta,
  Time = "Time",
  Conc = "Conc",
  Treatment = "Treatment",
  Reference = "R",
  Test = "T",
  method = "Linear-Up/Log-Down",
  Trt.cols = TRUE,
  details = FALSE,
  plot = TRUE
)
}
\arguments{
\item{dta}{Dataframe with concentration data from Test and Reference products.}

\item{Time}{Name of the column with time data.}

\item{Conc}{Name of the column with concentration data
(only for stacked data, \code{Trt.cols = FALSE}).}

\item{Treatment}{Name of the column with treatment/formulation information
(only for stacked data, \code{Trt.cols = FALSE}).}

\item{Reference}{Nomenclature in \code{dta} for the Reference product, defaults to \code{R}.}

\item{Test}{Nomenclature in \code{dta} for the Test product, defaults to \code{T}.}

\item{method}{Method for the AUC calculation:
\itemize{
\item \code{Linear-Up/Log-Down} (default method): Linear Trapezoidal Rule for for Increasing Values, Log Trapezoidal Rule for Decreasing Values.
\item \code{Linear/Log}: Linear Trapezoidal Rule for for Values below Cmax, Log Trapezoidal Rule for Values above Cmax.
\item \code{Linear}: Linear Trapezoidal Rule.
}}

\item{Trt.cols}{A logical value indicating whether treatment is presented in
columns/pivoted (\code{TRUE}), or in rows/stacked (\code{FALSE}).}

\item{details}{A logical value indicating whether detailed results will be
presented (\code{TRUE}), or if the function only returns the f2 factor
(\code{FALSE}). Defaults to \code{FALSE}.}

\item{plot}{A logical value indicating whether graphical representation will be
returned. Defaults to \code{TRUE}.}
}
\value{
Returns a list with the following elements:
\itemize{
\item Raw Concentration Data: Dataframe of input concentration data, with
treatment information in columns (pivoted).
\item Cumulative AUC: Dataframe with cumulative AUC over time, for Test and Reference product.
\item Reference AUClast: Vector of Reference AUClast.
\item Normalized AUC: dataframe with normalize Test and Reference AUC
concentrations over time, by Reference AUClast, until Tlast.
\item AUC f2 Factor: dataframe from \code{f2()} function.
}
}
\description{
\code{f2.AUC()} calculates the f2 Similarity factor for AUC, from concentration data.
}
\examples{
# Calculate f2 Factor for AUC
# when treatment data is pivoted:
f2.AUC(dta, Time = 'Time', Reference = 'R', Test = 'T')

# when treatment data is stacked:
f2.AUC(dta, Time = 'Time', Conc = 'Conc',
       Treatment = 'Treatment', Reference = 'R', Test = 'T')

}
\references{
Henriques, S.C.; Albuquerque, J.; Paixão, P.; Almeida, L.; Silva, N.E. (2023).
Alternative Analysis Approaches for the Assessment of Pilot Bioavailability/Bioequivalence Studies.
\emph{Pharmaceutics}. \emph{15}(5), 1430.
\href{https://doi.org/10.3390/pharmaceutics15051430}{10.3390/pharmaceutics15051430}.
}
\author{
Sara Carolina Henriques
}
