% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2.Cmax.R
\name{f2.Cmax}
\alias{f2.Cmax}
\title{Cmax Similarity f2 Factor}
\usage{
f2.Cmax(
  dta,
  Time = "Time",
  Conc = "Conc",
  Trt = "Treatment",
  Ref = "R",
  Test = "T",
  Trt.cols = TRUE,
  details = FALSE,
  plot = TRUE,
  col.R = "black",
  col.T = "blue"
)
}
\arguments{
\item{dta}{Dataframe with concentration data from Test and Reference products.}

\item{Time}{Name of the column with time data.}

\item{Conc}{Name of the column with concentration data
(only for stacked data, \code{Trt.cols = FALSE}).}

\item{Trt}{Name of the column with treatment/formulation information
(only for stacked data, \code{Trt.cols = FALSE}).}

\item{Ref}{Nomenclature in \code{dta} for the Reference product, defaults to \code{R}.}

\item{Test}{Nomenclature in \code{dta} for the Test product, defaults to \code{T}.}

\item{Trt.cols}{A logical value indicating whether treatment is presented in
columns/pivoted (\code{TRUE}), or in rows/stacked (\code{FALSE}).}

\item{details}{A logical value indicating whether detailed results will be
presented (\code{TRUE}), or if the function only returns the \eqn{f_2} factor
(\code{FALSE}). Defaults to \code{FALSE}.}

\item{plot}{A logical value indicating whether graphical representation will be
returned. Defaults to \code{TRUE}.}

\item{col.R}{Desired colour for plotting Reference data.
Defaults to \code{black}. Only used if \code{plot} is set to \code{TRUE}.}

\item{col.T}{Desired colour for plotting Test data.
Defaults to \code{blue}. Only used if \code{plot} is set to \code{TRUE}.}
}
\value{
Returns a list with the following elements:
\itemize{
\item \verb{Raw Concentration Data}: Dataframe of input concentration data, with
treatment information in columns (pivoted).
\item \verb{Reference Tmax}: Vector of Reference \eqn{t_{\text{max}}}
\item \verb{Reference Cmax}: Vector of Reference \eqn{C_{\text{max}}}
\item \verb{Normalized Concentrations}: dataframe with normalize Test and Reference
concentrations over time, by Reference \eqn{C_{\text{max}}}, until Reference \eqn{t_{\text{max}}}
\item \verb{Cmax f2 Factor}: dataframe from \code{\link[=f2]{f2()}} function.
}
}
\description{
\code{f2.Cmax()} calculates the \eqn{f_2} Similarity factor for \eqn{C_{\text{max}}},
from concentration data.
}
\details{
Henriques \emph{et al} (2023) proposed to used the \eqn{f_2} factor to assess the similarity on
the rate of drug absorption by normalizing Test and Reference mean concentration-time profiles to the
maximum plasma concentration (\eqn{C_{\text{max}}}) derived from the mean Reference profile, until
Reference \eqn{C_{\text{max}}} is observed (Reference \eqn{t_{\text{max}}}):
\deqn{
  C_{t}^{N} = 100 \cdot \frac{\overline{C}_{t}}{C_{\text{max R}}}, \text{ where } 0 \leq t \leq
  t_{\text{max R}}
}
where \eqn{C_{t}^{N}} is the normalized concentration at time \eqn{t}, \eqn{\overline{C}_{t}}
is the mean (Test or Reference) concentration at time \eqn{t}, \eqn{C_{\text{max R}}} is the
\eqn{C_{\text{max}}} of the Reference mean concentration-time profile, and \eqn{t_{\text{max R}}}
the time of observation of \eqn{C_{\text{max R}}}. The similarity \eqn{f_2} factor is calculated as
\deqn{
  C_{\text{max}} {f_2} = 50 \cdot \log \left( 100 \cdot \left[ 1 + \frac{1}{n} \sum_{t=1}^{t=n}
  \left(\overline{R}_t^N - \overline{T}_t^N\right)^2 \right]^{-0.5} \right)
}
where \eqn{C_{\text{max}} {f_2}} is the similarity factor calculated for \eqn{C_{\text{max}}},
\eqn{n} is the number of time points until Reference \eqn{t_{\text{max}}}, and \eqn{{\overline{R}_t^N}}
and \eqn{{\overline{T}_t^N}} are the are the normalized concentration at time \eqn{t} for Reference and
Test products, respectively.
}
\examples{
# For a dataframe with concentration-time data for Test and Reference product
# with treatment information pivoted:
dta_piv <- data.frame(
  Time = c(0, 0.25, 0.5, 0.75, 1, 1.5, 1.75, 2, 2.25, 2.5,
           2.75, 3, 3.25, 3.5, 3.75, 4, 6, 8, 12, 24),
  Ref = c(0.00, 221.23, 377.19, 494.73, 555.74,
          623.86, 615.45, 663.38, 660.29, 621.71,
          650.33, 622.28, 626.72, 574.94, 610.51,
          554.02, 409.14, 299.76, 162.85, 27.01),
  Test = c(0.00, 149.24, 253.05, 354.49, 412.49,
           530.07, 539.68, 566.30, 573.54, 598.33,
           612.63, 567.48, 561.10, 564.47, 541.50,
           536.92, 440.32, 338.78, 185.03, 31.13)
)
# Cmax f2 factor can be calculated as:
f2.Cmax(dta_piv, Time = 'Time', Ref = 'Ref', Test = 'Test')

# For a dataframe with concentration-time data for Test and Reference product
# with treatment information stacked:
dta_stk <- data.frame(
  Time = rep(c(0, 0.25, 0.5, 0.75, 1, 1.5, 1.75, 2, 2.25, 2.5,
               2.75, 3, 3.25, 3.5, 3.75, 4, 6, 8, 12, 24),2),
  Trt = c(rep('R',20), rep('T',20)),
  Conc = c(c(0.00, 221.23, 377.19, 494.73, 555.74,
             623.86, 615.45, 663.38, 660.29, 621.71,
             650.33, 622.28, 626.72, 574.94, 610.51,
             554.02, 409.14, 299.76, 162.85, 27.01),
           c(0.00, 149.24, 253.05, 354.49, 412.49,
             530.07, 539.68, 566.30, 573.54, 598.33,
             612.63, 567.48, 561.10, 564.47, 541.50,
             536.92, 440.32, 338.78, 185.03, 31.13))
)
# Cmax f2 factor can be calculated as:
f2.Cmax(dta_stk, Time = 'Time', Conc = 'Conc',
        Trt = 'Trt', Ref = 'R', Test = 'T', Trt.cols = FALSE)

}
\references{
Henriques, S.C.; Albuquerque, J.; Paixão, P.; Almeida, L.; Silva, N.E. (2023).
Alternative Analysis Approaches for the Assessment of Pilot Bioavailability/Bioequivalence Studies.
\emph{Pharmaceutics}. \emph{15}(5), 1430.
\href{https://doi.org/10.3390/pharmaceutics15051430}{10.3390/pharmaceutics15051430}.

Henriques, S.C.; Paixão, P.; Almeida, L.; Silva, N.E. (2023).
Predictive Potential of \eqn{C_{\text{max}}} Bioequivalence in Pilot Bioavailability/Bioequivalence
Studies, through the Alternative \eqn{f_2} Similarity Factor Method. \emph{Pharmaceutics}. \emph{15}(10), 2498.
\href{https://doi.org/10.3390/pharmaceutics15102498}{10.3390/pharmaceutics15102498}.
}
\author{
Sara Carolina Henriques
}
