% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{Geometric Mean}
\usage{
geomean(x, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be stripped
before the computation proceeds. Defaults to \code{FALSE}.}
}
\value{
Value of the geometric mean.
}
\description{
\code{geomean()} calculates geometric mean, i.e., the \emph{N}th root of the product of the \emph{N}
observations, equivalent to \code{exp(mean(log(x)))}.
}
\examples{
x <- c(49.66, 53.62, 52.25, 50.44, 46.84, 50.51, 47.26, 48.90, 49.83, 49.67)
geomean(x)

}
